Dropped LFS_F_COMPACT flags from lfsr_format
The argument for this flag is pretty brittle. Yes it's _technically_
possible to end up with a compactable filesystem during lfsr_format, but
it's pretty unlikely. And keeping LFS_F_COMPACT around means we'd always
need the lfsr_mtree_gc circuitry in lfsr_format, for such a niche
situation, that can be easily cleaned up in lfsr_mount.

So dropping for now.

No code changes, but this does mean one less feature to support:

           code          stack          ctx
  before: 37804           2608          620
  after:  37804 (+0.0%)   2608 (+0.0%)  620 (+0.0%)