Minor DEBUG/WARN/ERROR cleanup
Indention has changed quite a bit for most of these, so I figured it's a
good time to go back through and clean things up.

Also tried to adopt indention rules a bit better, adding indention after
string continuation. This was adopted in the scripts but not in the
actual code.