Moved gcksumdelta append into lfsr_mdir_commit__
This tweaks lfsr_rbyd_appendcksum_ to take the canonical cksum as an
argument, allowing the caller to append any non-tree tags they'd like
before finalizing the commit.

This lets us calculate our gcksumdelta directly in lfsr_mdir_commit__,
which simplifies things a bit and avoids a redundant cksum copy.

The code cost mostly cancels out due to the extra gcksumdelta append
needed in lfsr_formatinited, but it's worth it for the code
simplification:

           code          stack          ctx
  before: 38664           2648          644
  after:  38668 (+0.0%)   2648 (+0.0%)  644 (+0.0%)