rbyd: Fixed ambiguous stitching alt fix
What a funny bug.

I was wondering how jump ended up == branch, it turns out this was a
simple typo in the diverged-stitching rework.

What's extra humorous is I think the original code was trying to set
branch_ = jump, which wouldn't have worked. So the typo actually got
diverged-stitching to a more workable state than if there was no typo.

It just then broke the y-split logic.

Not randomly changing the branch when stitching is the correct fix.

---

This doesn't actually reduce any test failures since it's replacing the
previous jump >= branch "fix", which was really just a bandaid:

  test_rbyd+balance before: 1671/385878 failed
  test_rbyd+balance after:  1671/385878 failed (+0.0%)

Saves one ld/st though:

           code          stack          ctx
  before: 38584           2624          640
  after:  38580 (-0.0%)   2624 (+0.0%)  640 (+0.0%)