rbyd: Prioritized y-split prunes, fixed remaining balance issues!
The issue with relying solely on the jump > branch hack to disambiguate
recolorable prunes is that suddenly we're back in ambiguous territory
when we encounter post-split yellow nodes. We need to prioritize yellow
prunes or else post-split weights can become ambiguous.

This was the whole reason we added a 3rd color!

Fortunately we can still prioritize yellow nodes by trying both
recolorable prunes first, and only falling back to non-recolorable
prunes if we still have unreachable alts.

Ends up with a bit of code duplication, but gets things working again.

---

Turns out this was all we needed to get our rbyd operations perfectly
balanced! Now all test_rbyd tests are passing:

  test_rbyd+balance before: 1671/385878 failed
  test_rbyd+balance after:     0/385878 failed (-100.0%)

And after running the full test suite, can confirm _all_ tests are
passing with LFS_ASSERTRBYDBALANCE. So I think we have some pretty
decent confidence our rbyd algorithm maintains balance, even with range
removals:

  test+balance before: 20614/631541 failed
  test+balance after:      0/631541 failed (-100.0%)

It was an open question if this was even possible, so it's nice to see
some evidence balanced range removals are not a problem.

---

The current implementation is a bit hacky, so we do take a hit to code
size. Though we may be able to claw this back after cleaning things up:

           code          stack          ctx
  before: 38580           2624          640
  after:  38784 (+0.5%)   2624 (+0.0%)  640 (+0.0%)