Rearranged lfsr_mdir_commit prologue/epilogue a bit
Mainly to make our bshrub-in-mdir assert stricter (we need to update
mdirs first to assert bshrubs end up in the right mdirs), but also to
make lfsr_mdir_commit epilogue mirror the prologue a bit better. In this
function especially, readability is important.

And hey, code changes no worse for the wear:

           code          stack          ctx
  before: 36424           2608          640
  after:  36432 (+0.0%)   2608 (+0.0%)  640 (+0.0%)