Adopted lazy encoding for branch attrs
- LFSR_TAG_BRANCH      -+-> lfsr_data_frombranch
- LFSR_TAG_SHRUBBRANCH -'

This was a bit more involved than the others, since it requires changing
what we store in lfsr_bscratch_t.

I did poke around a bit with trying to reduce the total number of rbyd
allocations in lfsr_btree_commit__, but unfortunately we need both the
sibling rbyds and attr rbyds allocated at the same time while we
recursively figure out how to split/merge/commit.

I also almost forgot to include the dsize estimate, but fortunately this
was caught by our tests. We _do_ need branch dsizes, since they
contribute to any non-inlined bshrub's shrub estimate!

---

This was unfortunately a net-negative. We were actually getting a lot of
value from the smaller encoded branch size:

           code          stack          ctx
  before: 35472           2440          636
  after:  35508 (+0.1%)   2472 (+1.3%)  636 (+0.0%)

It may be worth reverting this, but I want to see how it interacts when
all of the existing eager rattr encoding logic has been removed.