Reverted lazy encoding for branch attrs
This was the one lazy attr that did _not_ save us code or stack, and in
fact hurt us a bit. Fortunately our lazy attr scheme still allows for
eager attr encoding, so we can revert just this one attr.

The reason is because we need to keep the children branches around as we
recursively commit up the btree, and, even assuming worst-case, the
branch encoding takes up way less RAM than an active rbyd.

Compare for yourself:

  in-RAM rbyd:         on-disk branch:
  .---+---+---+---.    .---+- -+- -+- -+- -.
  |     weight    |    | block             |
  +---+---+---+---+    +---+- -+- -+- -+- -'
  |     blocks    |    | trunk         |
  +               +    +---+- -+- -+- -+
  |               |    |     cksum     |
  +---+---+---+---+    '---+---+---+---'
  |s|   trunk     |
  +---+---+---+---+
  |p|    eoff     |
  +---+---+---+---+
  |     cksum     |
  '---+---+---+---'
  '-------.-------'    '---------.---------'
       24 bytes               13 bytes

Note we also don't have to care about alignment issues when passing
around the raw encoding.

There may also be something going on with the compiler assuming all
lfsr_rbyd_t pointers may alias, but it's a bit hard to tell.

Saves both code and stack:

           code          stack          ctx
  before: 35592           2472          636
  after:  35552 (-0.1%)   2440 (-1.3%)  636 (+0.0%)