Fix off by one in block boundary check
It was noticed that when copying files to the flash then eventually the fs gets corrupted.

After this finding a stress testing was implemented and additional traces added to littlefs.
With the testing it was noticed the off by one issue in a block boundary check.

This commit fixes the off by one check and all corruption traces were fixed.
The filesystem does not end up unmountable (corrupted) anymore after this fix.